# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My Calculator Project",
        "description": "My User Project Application API",
        "license": {
            "name": "IOD",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
        {
            "name": "Calculator",
            "description": "API for Calculus in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],

    "paths": {
        "/calculator/add": {
          "get": {
            "tags": [
              "Addition Service"
            ],
            "summary": "Add 2 numbers",
            "parameters": [
              {
                "name": "num1",
                "in": "query",
                "description": "the First Number"
              },
              {
                "name": "num2",
                "in": "query",
                "description": "the Second number"
              }
            ],
            "responses": {
              "200": {
                "description": "This service allows you to add two numbers"
              }
            }
          }
        },
        "/calculator/subtract": {
          "get": {
            "tags": [
              "Subtract Service"
            ],
            "summary": "Subtracts 2 numbers",
            "parameters": [
              {
                "name": "num1",
                "in": "query",
                "description": "the First Number"
              },
              {
                "name": "num2",
                "in": "query",
                "description": "the Second number"
              }
            ],
            "responses": {
              "200": {
                "description": "This service allows you to subtract two numbers"
              }
            }
          }
        },
        "/calculator/divide": {
          "get": {
            "tags": [
              "Dividing Service"
            ],
            "summary": "Divides 2 numbers",
            "parameters": [
              {
                "name": "num1",
                "in": "query",
                "description": "the First Number"
              },
              {
                "name": "num2",
                "in": "query",
                "description": "the Second number"
              }
            ],
            "responses": {
              "200": {
                "description": "This service allows you to divide two numbers"
              }
            }
          }
        },
        "/calculator/multiply": {
          "get": {
            "tags": [
              "Multiplication Service"
            ],
            "summary": "Multiply 2 numbers",
            "parameters": [
              {
                "name": "num1",
                "in": "query",
                "description": "the First Number"
              },
              {
                "name": "num2",
                "in": "query",
                "description": "the Second number"
              }
            ],
            "responses": {
              "200": {
                "description": "This service allows you to multiply two numbers"
              }
            }
          }
        }
      }
}